+++ src/extension.ts
@@ -4,18 +4,26 @@ import * as vscode from 'vscode';
 
 import { handleChangeEvent, handleFilesCreated, handleFilesDeleted, handleFilesRenamed, handlePastedFile } from "./event_handler";
 import { handleBuffer } from "./buffer_handler";
-import { initCodeSync } from "./utils/common";
-
+import { setupCodeSync, showConnectRepoView, showLogIn } from "./utils/setup_utils";
+import { COMMAND } from './constants';
+import { unSyncHandler, SignUpHandler, SyncHandler } from './commands_handler';
+    
 
 export async function activate(context: vscode.ExtensionContext) {
 	// Get the active text editor
 	const editor = vscode.window.activeTextEditor;
 	const repoName = vscode.workspace.name;
 	const repoPath = vscode.workspace.rootPath;
-
 	if (!repoPath || !repoName || !editor) { return; }
-	
-	await initCodeSync(repoPath);
+
+	vscode.commands.executeCommand('setContext', 'showLogIn', showLogIn());
+	vscode.commands.executeCommand('setContext', 'showConnectRepoView', showConnectRepoView(repoPath));
+
+	context.subscriptions.push(vscode.commands.registerCommand(COMMAND.triggerSignUp, SignUpHandler));
+	context.subscriptions.push(vscode.commands.registerCommand(COMMAND.triggerSync, SyncHandler));
+	context.subscriptions.push(vscode.commands.registerCommand(COMMAND.triggerUnsync, unSyncHandler));
+
+	await setupCodeSync(repoPath);
 
 	console.log(`Configured repo: ${repoPath}`);